# gateway/lab-exchange-api-spec.json
{
    "openapi": "3.0.0",
    "info": {
      "title": "Lab Exchange API",
      "version": "1.0.0"
    },
    "servers": [
      {
        "url": "https://api.prod.labdao.xyz/"
      },
      {
        "url": "https://api.stg.labdao.xyz/"
      }
    ],
    "paths": {
...
      "/flows": {
        "post": {
          "operationId": "addFlow",
          "summary": "Add Flow",
          "requestBody": {
            "required": true,
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "walletAddress": {
                      "type": "string",
                      "example": "0x1234567890abcdef1234567890abcdef12345678"
                    },
                    "toolCid": {
                      "type": "string",
                      "example": "Qm..."
                    },
                    "scatteringMethod": {
                      "type": "string"
                    },
                    "name": {
                      "type": "string"
                    },
                    "kwargs": {
                      "type": "object"
                    }
                  },
                  "required": [
                    "walletAddress",
                    "toolCid",
                    "scatteringMethod",
                    "name",
                    "kwargs"
                  ]
                }
              }
            }
          },
          "responses": {
            "201": {
              "description": "Flow created successfully",
              "content": {
                "application/json": {
                  "schema": {
                    "type": "object",
                    "properties": {
                      "cid": {
                        "type": "string"
                      }
                    }
                  }
                }
              }
            },
            "400": {
              "description": "Bad request, such as missing fields or invalid data"
            },
            "409": {
              "description": "Conflict, flow with the same CID already exists"
            },
            "500": {
              "description": "Internal Server Error"
            }
          }
        },
        "get": {
          "operationId": "listFlows",
          "summary": "List Flows",
          "parameters": [
            {
              "name": "cid",
              "in": "query",
              "required": false,
              "schema": {
                "type": "string"
              }
            },
            {
              "name": "name",
              "in": "query",
              "required": false,
              "schema": {
                "type": "string"
              }
            },
            {
              "name": "walletAddress",
              "in": "query",
              "required": false,
              "schema": {
                "type": "string"
              }
            }
          ],
          "responses": {
            "200": {
              "description": "List of flows",
              "content": {
                "application/json": {
                  "schema": {
                    "type": "array",
                    "items": {
                      "$ref": "#/components/schemas/Flow"
                    }
                  }
                }
              }
            },
            "500": {
              "description": "Internal Server Error"
            }
          }
        }
      },
      "/flows/{cid}": {
        "get": {
          "operationId": "getFlowByCid",
          "summary": "Get Flow",
          "parameters": [
            {
              "name": "cid",
              "in": "path",
              "required": true,
              "schema": {
                "type": "string"
              }
            }
          ],
          "responses": {
            "200": {
              "description": "Flow data",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/Flow"
                  }
                }
              }
            },
            "500": {
              "description": "Internal Server Error"
            }
          }
        },
        "patch": {
          "operationId": "updateFlow",
          "summary": "Update Flow",
          "parameters": [
            {
              "name": "cid",
              "in": "path",
              "required": true,
              "schema": {
                "type": "string"
              }
            }
          ],
          "responses": {
            "200": {
              "description": "Flow updated successfully",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/Flow"
                  }
                }
              }
            },
            "500": {
              "description": "Internal Server Error"
            }
          }
        }
      },
...
    },
    "components": {
      "schemas": {
...
        "Tool": {
          "type": "object",
          "properties": {
            "cid": {
              "type": "string"
            },
            "name": {
              "type": "string"
            },
            "walletAddress": {
              "type": "string"
            },
            "toolJson": {
              "type": "object"
            }
          }
        },
        "DataFile": {
          "type": "object",
          "properties": {
            "cid": {
              "type": "string"
            },
            "walletAddress": {
              "type": "string"
            },
            "filename": {
              "type": "string"
            },
            "jobInputs": {
              "type": "array",
              "items": {
                "$ref": "#/components/schemas/Job"
              }
            },
            "jobOutputs": {
              "type": "array",
              "items": {
                "$ref": "#/components/schemas/Job"
              }
            },
            "tags": {
              "type": "array",
              "items": {
                "$ref": "#/components/schemas/Tag"
              }
            },
            "timestamp": {
              "type": "string",
              "format": "date-time"
            }
          }
        },
        "Flow": {
          "type": "object",
          "properties": {
            "cid": {
              "type": "string"
            },
            "jobs": {
              "type": "array",
              "items": {
                "$ref": "#/components/schemas/Job"
              }
            },
            "name": {
              "type": "string"
            },
            "walletAddress": {
              "type": "string"
            }
          }
        },
        "Job": {
          "type": "object",
          "properties": {
            "bacalhauJobID": {
              "type": "string"
            },
            "state": {
              "type": "string"
            },
            "error": {
              "type": "string"
            },
            "walletAddress": {
              "type": "string"
            },
            "toolID": {
              "type": "string"
            },
            "tool": {
              "$ref": "#/components/schemas/Tool"
            },
            "flowID": {
              "type": "string"
            },
            "flow": {
              "$ref": "#/components/schemas/Flow"
            },
            "inputs": {
              "type": "array",
              "items": {
                "$ref": "#/components/schemas/DataFile"
              }
            },
            "outputs": {
              "type": "array",
              "items": {
                "$ref": "#/components/schemas/DataFile"
              }
            }
          }
        },
        "Tag": {
          "type": "object",
          "properties": {
            "name": {
              "type": "string"
            },
            "type": {
              "type": "string"
            },
            "dataFiles": {
              "type": "array",
              "items": {
                "$ref": "#/components/schemas/DataFile"
              }
            }
          }
        }
      }
    }
  }